Fixed make distcheck problem for occshared, occstatic and cubitshared builds, all passed; Minor changes to remove some warnings; Added cgm_version.h file and removed this part of definition from GeomeryQueryTool.
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/cgm/trunk@6096 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ geom/facet/FacetAttrib.cpp
@@ -72,9 +72,9 @@ CubitSimpleAttrib FacetAttrib::get_CSA() const
 bool FacetAttrib::equals( const CubitSimpleAttrib& csa ) const
 {
   // compare counts
-  if( csa.int_data_list().size() != numIntegers ||
-      csa.double_data_list().size() != numDoubles ||
-      csa.string_data_list().size() != numStrings )
+  if( csa.int_data_list().size() != (int)numIntegers ||
+      csa.double_data_list().size() != (int)numDoubles ||
+      csa.string_data_list().size() != (int)numStrings )
     return false;
 
     // compare strings first because most likely the
